[Decomp][Player] Cohesion split: extract battleground/arena methods to PlayerBattleGround.cpp
